requirements.txt
# Python Dependencies
# Recruiter ICP Job Tracker

# Core dependencies
python-dotenv==1.0.0
openai==1.12.0
requests==2.31.0

# Web scraping
beautifulsoup4==4.12.3
markdownify==0.11.6
playwright==1.41.2

# Apify
apify-client==1.6.3

# Exa API
exa-py==1.0.9

# Supabase
supabase==2.9.1

# Utilities
argparse>=1.4.0
pathlib>=1.0.1
